pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orix"
authors = [{name = "orix developers"}]
description = "orix is an open-source Python library for handling crystal orientation mapping data"
license = {file = "LICENSE"}
readme = {file = "README.rst", content-type = "text/x-rst"}
dynamic = ["version"]
requires-python = ">= 3.10"
classifiers=[
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "dask[array]",
    "diffpy.structure       >= 3.0.2",
    "h5py",
    "lazy_loader",
    "matplotlib             >= 3.6.1",
    "numba",
    "numpy",
    "pooch                  >= 0.13",
    # TODO: Remove once diffpy.structure >= 3.2.1
    "pycifrw",
    "scipy",
    "tqdm",
]

[project.optional-dependencies]
all = [  # NB! Update constants.py if this list is updated!
    "matplotlib-scalebar",
    "numpy-quaternion",
]
doc = [
    "ipykernel",  # Used by nbsphinx to execute notebooks
    "memory_profiler",
    "nbconvert                      >= 7.16.4",
    "nbsphinx                       >= 0.7",
    # Restriction due to https://github.com/pyxem/orix/issues/570
    "numpydoc != 1.9.0",
    "pydata-sphinx-theme",
    "scikit-image",
    "scikit-learn",
    "sphinx                         >= 3.0.2",
    "sphinx-codeautolink[ipython]",
    "sphinx-copybutton              >= 0.2.5",
    "sphinx-design",
    "sphinx-gallery",
    "sphinxcontrib-bibtex           >= 1.0",
]
tests = [
    "numpydoc",
    "pytest                         >= 5.4",
    # TODO: Remove restriction once a new release fixing
    # https://github.com/pytest-dev/pytest-rerunfailures/issues/302 is available.
    "pytest-rerunfailures           < 16.0",
    "pytest-xdist",
]
coverage = [
    "coverage                       >= 5.0",
    "pytest-cov                     >= 2.8.1",
]
dev = [
    "black[jupyter]",
    "hatch",
    "isort                          >= 5.10",
    "licenseheaders",
    "outdated",
    "pre-commit                     >= 1.16",
    "orix[doc,tests,coverage]",
]

[tool.hatch.version]
path = "orix/__init__.py"

[tool.coverage.report]
precision = 2

[tool.coverage.run]
branch = false
source = ["orix"]
relative_files = true
omit = [
    "orix/__init__.py",
    "orix/tests/**/*.py",
]

[tool.pytest.ini_options]
addopts = [
    "-ra",
    "--ignore=doc/_static/img/colormap_banners/create_colormap_banners.py",
    "--ignore=examples/*/*.py",
    "--strict-markers",
]
doctest_optionflags = "NORMALIZE_WHITESPACE"
filterwarnings = [
    "ignore:Deprecated call to `pkg_resources:DeprecationWarning",
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning",
]
markers = [
    "slow: mark test as slow",
]

[tool.isort]
profile = "black"
filter_files = true
force_sort_within_sections = true
